feat: HeroSection 컴포넌트를 URL의 pathParams를 사용하여 API에서 데이터를 가져오도록 수정
